+++ export/src/main/java/tech/ydb/export/result/ExportToS3Result.java
@@ -1,0 +1,10 @@
+package tech.ydb.export.result;
+
+import tech.ydb.export.YdbExport;
+
+public class ExportToS3Result {
+
+    public ExportToS3Result(YdbExport.ExportToS3Result exportToS3Result) {
+        // do nothing
+    }
+}